+++ x/group/internal/math/dec.go
@@ -46,6 +46,10 @@ func (x Dec) IsPositive() bool {
 	return !x.dec.Negative && !x.dec.IsZero()
 }
 
+func (x Dec) IsFinite() bool {
+	return x.dec.Form != apd.Finite
+}
+
 // NewDecFromString returns a new Dec from a string
 // It only support finite numbers, not NaN, +Inf, -Inf
 func NewDecFromString(s string) (Dec, error) {
@@ -55,7 +59,7 @@ func NewDecFromString(s string) (Dec, error) {
 	}
 
 	if d.Form != apd.Finite {
-		return Dec{}, grouperrors.ErrInvalidDecString.Wrapf("expected a finite decimal, got %s", s)
+		return Dec{}, errors.ErrInvalidDecString.Wrapf("expected a finite decimal, got %s", s)
 	}
 
 	return Dec{*d}, nil