+++ x/group/module/abci.go
@@ -1,0 +1,16 @@
+package module
+
+import (
+	sdk "github.com/cosmos/cosmos-sdk/types"
+	"github.com/cosmos/cosmos-sdk/x/group/keeper"
+)
+
+func EndBlocker(ctx sdk.Context, k keeper.Keeper) {
+	if err := k.TallyProposalsAtVPEnd(ctx); err != nil {
+		panic(err)
+	}
+
+	if err := k.PruneProposals(ctx); err != nil {
+		panic(err)
+	}
+}